refactor: 合并拓展类代码
