+++ src/BootstrapBlazor/Extensions/JSModuleExtensions.cs
@@ -47,6 +47,16 @@ public static class JSModuleExtensions
     }
 
     /// <summary>
+    /// 导入js模块
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <returns></returns>
+    private static async Task<IJSObjectReference> GetModule(IJSRuntime jsRuntime)
+    {
+        return await jsRuntime.InvokeAsync<IJSObjectReference>(identifier: "import", "./_content/BootstrapBlazor/modules/utility.js");
+    }
+
+    /// <summary>
     /// 在新标签页打开指定网址
     /// </summary>
     /// <param name="jsRuntime"></param>
@@ -54,7 +64,287 @@ public static class JSModuleExtensions
     /// <returns></returns>
     public static async ValueTask OpenBlankUrl(this IJSRuntime jsRuntime, string url)
     {
-        var jSObjectReference = await jsRuntime.InvokeAsync<IJSObjectReference>(identifier: "import", "./_content/BootstrapBlazor/modules/utility.js");
-        await jSObjectReference.InvokeVoidAsync("openBlank", url);
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("openBlank", url);
     }
+
+    /// <summary>
+    /// 获取是否为移动设备
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <returns></returns>
+    public static async Task<bool> GetIsMobileDevice(this IJSRuntime jsRuntime)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<bool>("isMobileDevice");
+    }
+
+    /// <summary>
+    /// 动态运行js代码
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="script"></param>
+    /// <returns></returns>
+    public static async Task<T> Eval<T>(this IJSRuntime jsRuntime, string script)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<T>("runEval", script);
+    }
+
+    /// <summary>
+    /// 动态运行js代码
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="script"></param>
+    /// <returns></returns>
+    public static async Task Eval(this IJSRuntime jsRuntime, string script)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("runEval", script);
+    }
+
+    /// <summary>
+    /// 动态运行js代码
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="script"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task Function(this IJSRuntime jsRuntime, string script, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("runFunc", script, args);
+    }
+
+    /// <summary>
+    /// 动态运行js代码
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    /// <param name="jsRuntime"></param>
+    /// <param name="script"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task<T> Function<T>(this IJSRuntime jsRuntime, string script, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<T>("runFunc", script, args);
+    }
+
+    /// <summary>
+    /// 清空浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    public static async Task ConsoleClear(this IJSRuntime jsRuntime)
+    {
+        await jsRuntime.InvokeVoidAsync("eval", "console.clear()");
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleLog(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "log", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleWarn(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "warn", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleError(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "error", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleInfo(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "info", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleAssert(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "assert", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleDir(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "dir", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleTime(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "time", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleTimeEnd(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "timeEnd", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleCount(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "count", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleGroup(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "group", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleGroupEnd(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "groupEnd", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleTable(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "table", args);
+    }
+
+    /// <summary>
+    /// 输出到浏览器控制台
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="args"></param>
+    /// <returns></returns>
+    public static async Task ConsoleTrace(this IJSRuntime jsRuntime, params object?[]? args)
+    {
+        var module = await GetModule(jsRuntime);
+        await module.InvokeVoidAsync("doConsole", "trace", args);
+    }
+
+    /// <summary>
+    /// 动态修改head标签
+    /// <para>
+    /// razor示例：
+    /// <code>
+    /// @inject <see cref="IJSRuntime"/> JSRuntime
+    /// 添加一个link标签
+    /// var result = await JSRuntime.ChangeMetaAsync(<see langword="true"/>,<see cref="HeadMetaType.Link"/>,"stylesheet","styles.css")
+    /// 移除一个link标签
+    /// var result = await JSRuntime.ChangeMetaAsync(<see langword="false"/>,<see cref="HeadMetaType.Link"/>,"stylesheet","styles.css")
+    /// </code>
+    /// </para>
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="isAdd"><see langword="true"/> 添加，<see langword="false"/> 移除</param>
+    /// <param name="headMetaType">head标签元素类型</param>
+    /// <param name="rel">类型</param>
+    /// <param name="href">地址</param>
+    /// <returns></returns>
+    public static async Task<bool> ChangeMetaAsync(this IJSRuntime jsRuntime, bool isAdd, HeadMetaType headMetaType, string rel, string href)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<bool>("changeMeta", isAdd, headMetaType.ToDescriptionString(), rel, href);
+    }
+
+    /// <summary>
+    /// 获取元素的指定属性
+    /// </summary>
+    /// <param name="jsRuntime"></param>
+    /// <param name="id"></param>
+    /// <param name="tag"></param>
+    /// <returns></returns>
+    public static async Task<T> GetElementProperties<T>(this IJSRuntime jsRuntime, string id, string tag)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<T>("getElementProperties", id, tag);
+    }
+
+    /// <summary>
+    /// 获取指定元素 CSS 的值
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
+    /// <param name="jsRuntime"></param>
+    /// <param name="id"></param>
+    /// <param name="propertyName"></param>
+    /// <returns></returns>
+    public static async Task<T?> GetCSSValue<T>(this IJSRuntime jsRuntime, string id, string propertyName)
+    {
+        var module = await GetModule(jsRuntime);
+        return await module.InvokeAsync<T?>("getCSSValue", id, propertyName);
+    }
+
 }
